+++ c/include/libsbp/system.h
@@ -1,6 +1,6 @@
 /*
  * Copyright (C) 2015-2018 Swift Navigation Inc.
- * Contact: Swift Navigation <dev@swiftnav.com>
+ * Contact: https://support.swiftnav.com
  *
  * This source is subject to the license found in the file 'LICENSE' which must
  * be be distributed together with this source. All other rights reserved.
@@ -525,7 +525,8 @@ typedef struct SBP_ATTR_PACKED {
   u8 flags;           /**< Status flags (reserved) */
   u8 n_group_msgs;    /**< Size of list group_msgs */
   u16 group_msgs[0];   /**< An inorder list of message types included in the Solution Group,
-including GROUP_META itself */
+including GROUP_META itself
+ */
 } msg_group_meta_t;