+++ haskell/src/SwiftNav/SBP/TH.hs
@@ -5,7 +5,7 @@
 -- Module:      SwiftNav.SBP.TH
 -- Copyright:   Copyright (C) 2015 Swift Navigation, Inc.
 -- License:     LGPL-3
--- Maintainer:  Mark Fine <dev@swiftnav.com>
+-- Contact:     https://support.swiftnav.com
 -- Stability:   experimental
 -- Portability: portable
 --