+++ python/sbp/client/drivers/network_drivers.py
@@ -1,5 +1,5 @@
 # Copyright (C) 2015 Swift Navigation Inc.
-# Contact: Bhaskar Mookerji <mookerji@swiftnav.com>
+# Contact: https://support.swiftnav.com
 #
 # This source is subject to the license found in the file 'LICENSE' which must
 # be be distributed together with this source. All other rights reserved.
@@ -84,6 +84,23 @@ class TCPDriver(BaseDriver):
                 continue
             break
 
+    def _perform_io(self, io_func, validate_data=lambda _data: True):
+        data = None
+        while True:
+            try:
+                data = io_func()
+            except socket.timeout as socket_error:
+                self._reconnect(socket_error)
+            except socket.error as socket_error:
+                # this is fine, just retry
+                if socket_error.errno == errno.EINTR:
+                    continue
+                self._reconnect(socket_error)
+            if not validate_data(data):
+                continue
+            break
+        return data
+
     def _read(self, size):
         """
         Read wrapper.
@@ -93,21 +110,16 @@ class TCPDriver(BaseDriver):
         size : int
           Number of bytes to read
         """
-        data = ''
-        while True:
-            try:
-                data = self.handle.recv(size)
-            except socket.timeout as socket_error:
-                self._reconnect(socket_error)
-            except socket.error as socket_error:
-                # this is fine, just retry
-                if socket_error.errno == errno.EINTR:
-                    continue
-                self._reconnect(IOError)
+
+        def read():
+            return self.handle.recv(size)
+
+        def validate_data(data):
             if not data:
                 self._reconnect(IOError)
-            break
-        return data
+            return bool(data)
+
+        return self._perform_io(read, validate_data)
 
     def flush(self):
         pass
@@ -121,12 +133,11 @@ class TCPDriver(BaseDriver):
         s : bytes
           Bytes to write
         """
+        def write():
+            return self.handle.sendall(s)
         try:
             self._write_lock.acquire()
-            self.handle.sendall(s)
-        except socket.timeout:
-            self._connect()
-        except socket.error:
-            raise IOError
+            count = self._perform_io(write)
         finally:
             self._write_lock.release()
+        return count